Support extra wide screens (18:9) on Android devices
Added "android:resizeableActivity=true" to Manifest so that device will start application in wide mode by default.
The documentation of this flag is confusing - it specifies that flag should be already TRUE
if application targeting API level 26, but it is not (at least system still does not start application wide by default).

This change in Manifest file requires using JAR file for android-24+ for building APK file.